No need to run device-plugin manually anymore.
It's an addon in GCP now. The manifest for that is here:
https://github.com/kubernetes/kubernetes/blob/master/cluster/addons/device-plugins/nvidia-gpu/daemonset.yaml

Instead of that a pause container is run, so that the installer can remain an init container.
Also, added node affinity to only run on nodes that have a label with key 'cloud.google.com/gke-accelerator'.

---

There are three files now:
1) device-plugin-daemonset.yaml
2) daemonset.yaml
3) daemonset-preloaded.yaml

(1) already existed but its name no longer makes sense. (2) is copy of (1).
(3) is copy of (2) except that the installer image is assumed to be present on
the node. (1) would be deleted once tests in kubernetes/kubernetes are updated
to point to (2) or (3).