Pin nvidia-driver-installer to specific version
Due to an update in the cos-gpu-installer image, the GPU OSS test started failing (see: https://github.com/kubernetes/kubernetes/issues/93187). The goal of the OSS test is not to test the driver installer itself, so I'd like to pin the driver installer version to a specific version, to avoid breaking tests when new driver installers are released.