Add Xid Config into gpuConfig
Add Xid Config into gpuConfig in nvidia_gpu.go file after parsingGPUConfig. 
The reason are following: in parsing GPUConfig function, we will reset gpuConfig when we get error after AddDefaultsAndValidate(). If we add AddXidConfig() inside the parsing GPUConfig function and get successfully passed in, but AddDefaultsAndValidate() return error. The gpuConfig will be reset to gpumanager.GPUConfig{}

So it is better to check the xid at the end.